Remove unused mapper modules
